Introduce simple macro wrappers for testing Lua
Avoids significant portion of the boilerplate and accounts for the
majority of checks.  Error on test failure now points to code, which is
a bit more convenient.

While at it (since lots of lines were rewritten anyway):
 - remove some unnecessary "\n" at the end
 - replace trailing space in string literals with explicit "\n" that's
   easier to notice
 - improve formatting
   * add some missing spaces
   * break statements into multiple lines less
   * improve alignment of string literals with Lua code

This also adds some more checks as now it's harder to forget about them.