Improve performance of comparing files by content
Simplify body of get_contents_fingerprint() down to a single fread()
followed by XXH3_64bits() call.  This gets rid of state management which
requires more memory and CPU resources.  The function originally
computed hash of the whole file.

The effect of this optimization shouldn't be huge, but an improvement
for large number of files is likely.  For 2 directories with 53003 files
each (directory and its backup), the measured userspace gain was a bit
over 5%, which isn't nothing.