Add strerror(errno) to individual syscalls errors
Individual error messages when 'syscalls' is on lacked system error
string which is present when errors are aggregated at the end of an
operation.

Re-organize the code to reuse the same formatting code in both cases so
they won't diverge in the future.

Thanks to an anonymous at Vifm Q2A site.

See
https://q2a.vifm.info/1756/avoiding-excessive-error-dialogs-during-copy-move?show=1767#c1767