Replace "--" in shell command without necessity
Commits c470960ab3abd5b9d2c70062d7bd1d6c51f6359e and
51ea56fee5f45e9e13a8e929201687bed8e04c77 add "--" to shell invocation
commands unconditionally, which turns out to break at least fish shell.

Unlike POSIX shells, fish's `-c` option accepts command as an argument
(POSIX shells accept command as a positional arguments and `-c` only
indicates that a command is coming).

This "fix" avoids breaking normal commands when fish is used, but the
end result is:
 - commands NOT starting with a dash WORK in POSIX shells
 - commands NOT starting with a dash WORK in fish shell
 - commands starting with a dash WORK in POSIX shells
 - commands starting with a dash DON'T WORK in fish shell
while prior to aforementioned commits the situation was as follows:
 - commands NOT starting with a dash WORK in POSIX shells
 - commands NOT starting with a dash WORK in fish shell
 - commands starting with a dash DON'T WORK in POSIX shells
 - commands starting with a dash WORK in fish shell

Ideally, all commands will work everywhere, but not sure how to do that
reliably and because commands starting with a dash are rare, I think
it's bearable if it's broken in a non-conforming shell.

Thanks to AERDU.

Fixes #994 on GitHub.